app/requirements.txt
pillow
numpy
scipy
scikit-image
flask
sklearn
simplejson